requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
django-redis==4.9.0
djangorestframework==3.8.2
Pillow==5.1.0
PyMySQL==0.8.1
pytz==2018.4
redis==2.10.6
